+++ manta-crypto/src/algebra/diffie_hellman.rs
@@ -123,8 +123,7 @@ where
     where
         W: Write,
     {
-        self.generator.encode(&mut writer)?;
-        Ok(())
+        self.generator.encode(&mut writer)
     }
 }