+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [\#259](https://github.com/Manta-Network/manta-rs/pull/257) Add server binaries for the trusted setup
 
 ### Changed
+- [\#274](https://github.com/Manta-Network/manta-rs/pull/274) Update TS client installer script to use release 0.5.6.
 
 ### Deprecated